+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 * Deprecates `Instabug.start` in favour of the new `Instabug.init` API.
+* Adds `hungarian` and `finnish` locales support
+* Adds missing mapping for `norwegian` and `slovak` locales on iOS
 
 ## 11.6.0 (2022-12-29)